A* finally returns a path!
It's probably incorrect, but at least it returns one. Next step is to
visually plot the path and see how it looks. Since we're not paying any
attention to cell contents yet, it should look like a diagonal line. It
probably doesn't.

TODO: If it doesn't, figure out why not and fix it.
TODO: If it does, replace FudgeTaxiDistance with PathChoice, which is
      the same except stepping onto an occupied cell has infinite cost.